+++ react/features/base/participants/reducer.ts
@@ -19,6 +19,7 @@ import {
 } from './actionTypes';
 import { LOCAL_PARTICIPANT_DEFAULT_ID, PARTICIPANT_ROLE } from './constants';
 import { isParticipantModerator } from './functions';
+import { LocalParticipant, Participant } from './types';
 
 /**
  * Participant object.
@@ -36,45 +37,6 @@ import { isParticipantModerator } from './functions';
  * {@code false}.
  * @property {string} email - Participant email.
  */
-
-export interface Participant {
-    avatarURL?: string;
-    botType?: string;
-    conference?: Object;
-    connectionStatus?: string;
-    dominantSpeaker?: boolean;
-    e2eeSupported?: boolean;
-    email?: string;
-    features?: {
-        'screen-sharing'?: boolean;
-    };
-    id: string;
-    isFakeParticipant?: boolean;
-    isJigasi?: boolean;
-    isLocalScreenShare?: boolean;
-    isReplacing?: number;
-    isVirtualScreenshareParticipant?: boolean;
-    loadableAvatarUrl?: string;
-    loadableAvatarUrlUseCORS?: boolean;
-    local?: boolean;
-    name?: string;
-    pinned?: boolean;
-    presence?: string;
-    raisedHandTimestamp?: number;
-    remoteControlSessionStatus?: boolean;
-    role?: string;
-    supportsRemoteControl?: boolean;
-}
-
-export interface LocalParticipant extends Participant {
-    audioOutputDeviceId?: string;
-    cameraDeviceId?: string;
-    micDeviceId?: string;
-    startWithAudioMuted?: boolean;
-    startWithVideoMuted?: boolean;
-    userSelectedMicDeviceId?: string;
-    userSelectedMicDeviceLabel?: string;
-}
 
 /**
  * The participant properties which cannot be updated through
