+++ react/features/lobby/reducer.ts
@@ -1,6 +1,6 @@
 // @ts-ignore
 import { CONFERENCE_JOINED, CONFERENCE_LEFT, SET_PASSWORD } from '../base/conference';
-import { Participant } from '../base/participants/reducer';
+import { Participant } from '../base/participants/types';
 import ReducerRegistry from '../base/redux/ReducerRegistry';
 
 import {
